organized the code, naming, fixes
